+++ src/main/kotlin/com/petqua/application/product/WishProductService.kt
@@ -1,19 +1,20 @@
-package com.petqua.application.wish
+package com.petqua.application.product
 
-import com.petqua.application.product.DecreaseWishCountEvent
+import com.petqua.application.product.dto.DeleteWishCommand
+import com.petqua.application.product.dto.SaveWishCommand
 import com.petqua.common.domain.existByIdOrThrow
 import com.petqua.common.domain.findByIdOrThrow
 import com.petqua.domain.member.MemberRepository
 import com.petqua.domain.product.ProductRepository
-import com.petqua.domain.wish.WishProductRepository
+import com.petqua.domain.product.WishProductRepository
 import com.petqua.exception.member.MemberException
 import com.petqua.exception.member.MemberExceptionType.NOT_FOUND_MEMBER
 import com.petqua.exception.product.ProductException
 import com.petqua.exception.product.ProductExceptionType.NOT_FOUND_PRODUCT
-import com.petqua.exception.wish.WishProductException
-import com.petqua.exception.wish.WishProductExceptionType.ALREADY_EXIST_WISH_PRODUCT
-import com.petqua.exception.wish.WishProductExceptionType.NOT_FOUND_WISH_PRODUCT
-import com.petqua.presentation.wish.WishProductResponse
+import com.petqua.exception.product.WishProductException
+import com.petqua.exception.product.WishProductExceptionType.ALREADY_EXIST_WISH_PRODUCT
+import com.petqua.exception.product.WishProductExceptionType.NOT_FOUND_WISH_PRODUCT
+import com.petqua.presentation.product.WishProductResponse
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -44,7 +45,6 @@ class WishProductService(
             wishProductRepository.findByIdOrThrow(command.wishProductId, WishProductException(NOT_FOUND_WISH_PRODUCT))
         wishProduct.validateOwner(command.memberId)
         wishProductRepository.delete(wishProduct)
-        publisher.publishEvent(DecreaseWishCountEvent(wishProduct.productId))
     }
 
     fun readAll(memberId: Long): List<WishProductResponse> {
