+++ src/main/kotlin/com/petqua/domain/product/WishProductCustomRepositoryImpl.kt
@@ -1,12 +1,11 @@
-package com.petqua.domain.wish
+package com.petqua.domain.product
 
 import com.linecorp.kotlinjdsl.dsl.jpql.jpql
 import com.linecorp.kotlinjdsl.render.jpql.JpqlRenderContext
 import com.linecorp.kotlinjdsl.render.jpql.JpqlRenderer
 import com.petqua.common.util.createQuery
-import com.petqua.domain.product.Product
 import com.petqua.domain.store.Store
-import com.petqua.presentation.wish.WishProductResponse
+import com.petqua.presentation.product.WishProductResponse
 import jakarta.persistence.EntityManager
 import org.springframework.stereotype.Repository
